refactor(training): merge L1-reg + L1-anchor + unfreeze into qft_direct_training
One driver, three subcommands (l1-reg / l1-anchor / unfreeze), each lifting
its source script's behavior verbatim with new 2_direct_training/ output
paths and the CUDA_DEVICE_ORDER=PCI_BUS_ID pin. Removes the three source
scripts (qft_identity_regularization.py, div2k_8q_l1_init_anchor.py,
qft_unfreeze.py). Behavior-preserving: all three modes smoke-verified to
reproduce the cell schema (l1-reg PSNRs match the prior run).
